+++ game/MAIN/MainInit.c
@@ -261,7 +261,7 @@ void MainInit_JitPoolsNew(struct GameTracker *gGT)
 	gGT->ptrRenderBucketInstance = MEMPACK_AllocMem(renderBucketSize);
 #else
 	// NOTE(aalhendi): Native reuses static RDATA scratch for existing PC memory headroom.
-	gGT->ptrRenderBucketInstance = (void *)((uintptr_t)&rdata.s_STATIC_GNORMALZ[0] + 148);
+	gGT->ptrRenderBucketInstance = (void *)((u32)&rdata.s_STATIC_GNORMALZ[0] + 148);
 #endif
 
 	for (int i = 0; i < 3; i++)
